+++ resource.h
@@ -122,6 +122,7 @@
 #define IDC_GSET_DURATION_DEBUG  2254   // "Duration debug" checkbox
 #define IDC_GSET_SHADOW_PARAMS      2255   // "Shadow VST3 params" checkbox
 #define IDC_GSET_CHUNK_ALL_INSTANT  2256   // "Chunk all on instant path" checkbox
+#define IDC_GSET_RECALL_LOG         2259   // "Write recall log to file" checkbox
 
 // ---- Scenes window dock toggle + marker option ----------------------------
 // (2145-2147 now used – see IDC_SNAPNAME etc. above)